# HG changeset patch # User Todd Jackson <tmjackso@uci.edu> # Date 1325139243 28800 # Node ID 7759b2eadd6406203f10be4337fc50372635a0f2 # Parent  a76429bc8861ffa31f7e24480129173d31e9e193 Refactor and fix PBKDF2 implementation for Skein

+++ include/llvm/MultiCompiler/MultiCompilerOptions.h
@@ -7,11 +7,12 @@
 
 #ifndef MULTI_COMPILER_OPTIONS_H
 #define MULTI_COMPILER_OPTIONS_H
+#include <stdint.h>
 #include <string>
 
 namespace multicompiler {
  extern unsigned int RandomStackLayout;
- extern unsigned int MultiCompilerSeed;
+ extern uint64_t MultiCompilerSeed;
  extern int PreRARandomizerRange;
  extern unsigned int MaxStackFramePadding;
  extern std::string RNGStateFile;
